move checkbox group examples
