+++ src/examples/src/widgets/checkbox-group/Basic.tsx
@@ -1,0 +1,25 @@
+import { create, tsx } from '@dojo/framework/core/vdom';
+import CheckboxGroup from '../../../../checkbox-group/index';
+import { icache } from '@dojo/framework/core/middleware/icache';
+
+const factory = create({ icache });
+
+const App = factory(function({ properties, middleware: { icache } }) {
+	const { get, set } = icache;
+
+	return (
+		<virtual>
+			<CheckboxGroup
+				label="pets"
+				name="standard"
+				options={[{ value: 'cat' }, { value: 'dog' }, { value: 'fish' }]}
+				onValue={(value) => {
+					set('standard', value);
+				}}
+			/>
+			<pre>{`${get('standard')}`}</pre>
+		</virtual>
+	);
+});
+
+export default App;
