+++ src/examples/src/widgets/checkbox-group/CustomLabel.tsx
@@ -1,0 +1,29 @@
+import { create, tsx } from '@dojo/framework/core/vdom';
+import CheckboxGroup from '../../../../checkbox-group/index';
+import { icache } from '@dojo/framework/core/middleware/icache';
+
+const factory = create({ icache });
+
+const App = factory(function({ properties, middleware: { icache } }) {
+	const { get, set } = icache;
+
+	return (
+		<virtual>
+			<CheckboxGroup
+				label="colours"
+				name="colours"
+				options={[
+					{ value: 'red', label: 'Rouge' },
+					{ value: 'green', label: 'Vert' },
+					{ value: 'blue', label: 'Bleu' }
+				]}
+				onValue={(value) => {
+					set('colours', value);
+				}}
+			/>
+			<pre>{`${get('colours')}`}</pre>
+		</virtual>
+	);
+});
+
+export default App;
