+++ src/examples/src/widgets/checkbox-group/InitialValue.tsx
@@ -1,0 +1,26 @@
+import { create, tsx } from '@dojo/framework/core/vdom';
+import CheckboxGroup from '../../../../checkbox-group/index';
+import { icache } from '@dojo/framework/core/middleware/icache';
+
+const factory = create({ icache });
+
+const App = factory(function({ properties, middleware: { icache } }) {
+	const { get, set } = icache;
+
+	return (
+		<virtual>
+			<CheckboxGroup
+				label="favourite names"
+				initialValue={['tom']}
+				name="initial-value"
+				options={[{ value: 'tom' }, { value: 'dick' }, { value: 'harry' }]}
+				onValue={(value) => {
+					set('initial-value', value);
+				}}
+			/>
+			<pre>{`${get('initial-value')}`}</pre>
+		</virtual>
+	);
+});
+
+export default App;
